requirements.txt
fastmcp>=0.1.0
google-cloud-bigquery>=3.0.0
